requirements.txt
aiocoap==0.4.1
APScheduler==3.7.0
Flask==2.1.0
Flask_RESTful==0.3.9
jsonschema==3.2.0
paho_mqtt==1.6.1
python-dotenv==0.21.0
requests==2.25.1

